daemon: fix not working motor, add motor test mode '-tm' evaluate Heater state in seperate Thread use Heater information in logic/return state each mode has it's own loop Shield Version in configfile press as int not double fix read string configs
middleware:
do not recalculate "remaining time", percent calculation always from time

install:
add backupConfig script #19

+++ daemon/hardwareFunctions.h
@@ -16,7 +16,7 @@ See GPLv3.htm in the main folder for details.
 #include "bool.h"
 
 double readTemp(struct Daemon_Values *dv);
-double readPress(struct Daemon_Values *dv);
+int32_t readPress(struct Daemon_Values *dv);
 bool HeatOn(struct Daemon_Values *dv);
 bool HeatOff(struct Daemon_Values *dv);
 void setMotorPWM(uint16_t pwm, struct Daemon_Values *dv);
